Rename options => init
